drawbot_code/requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.3
Shapely==1.7.1
six==1.16.0
svgwrite==1.4.1
tifffile==2021.4.8
tqdm==4.60.0
